+++ deployment/dockerfiles/origin-discovery
@@ -24,4 +24,4 @@ COPY --from=origin-js-build /app/origin-js/ /usr/local/lib/node_modules/origin
 RUN npm link origin
 RUN npm install --quiet --no-progress
 
-CMD ["node", "src/apollo/index.js"]
+CMD ["node", "src/apollo/app.js"]
